requirements.txt
# Core dependencies
python-dotenv==1.2.1
streamlit==1.55.0
pandas==2.3.3

# LangChain ecosystem
langchain==1.1.3
langchain-chroma==1.1.0
langchain-huggingface==1.2.0
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langchain-openai==1.1.7

# LangGraph
langgraph==1.0.10

# Anthropic
anthropic==0.76.0

# Vector store and embeddings
chromadb==1.4.1
sentence-transformers==5.2.0

# SSL (corporate proxy)
pip-system-certs==5.3
